render skills and radar from common source
