+++ src/components/skills.tsx
@@ -1,93 +1,43 @@
 import React from 'react';
 import { Card, Container, Row, Col } from 'react-bootstrap';
+import skills from '../data/skills';
+import { Skill } from '../types/domain';
+
+interface SkillCardProps {
+  skill: Skill;
+}
+
+const SkillCard = (props: SkillCardProps) => {
+  const skillProperties = ['category', 'level', 'url', 'description']
+  return (
+    <Card bg="light" style={{ width: '18rem' }}>
+      <Card.Header>{props.skill.title}</Card.Header>
+      <Card.Body>
+        <Card.Text>
+          <ul>
+            {skillProperties.map((key, index) => (
+              <li key={index}><b>{key}</b>: {props.skill[key]}</li>
+            ))}
+          </ul>
+        </Card.Text>
+      </Card.Body>
+    </Card>
+  );
+};
 
 const Skills = () => {
   return (
     <Container className="d-flex justify-content-center">
       <Row>
-        <Col>
-          <Card bg="light" style={{ width: '18rem' }}>
-            <Card.Header>JavaScript</Card.Header>
-            <Card.Body>
-              <Card.Text>
-                <ul>
-                  <li>ES6+</li>
-                  <li>Node.js</li>
-                  <li>React</li>
-                  <li>Angular</li>
-                  <li>Vue.js</li>
-                </ul>
-              </Card.Text>
-            </Card.Body>
-          </Card>
-        </Col>
-        <Col>
-          <Card bg="light" style={{ width: '18rem' }}>
-            <Card.Header>TypeScript</Card.Header>
-            <Card.Body>
-              <Card.Text>
-                <ul>
-                  <li>Interfaces</li>
-                  <li>Generics</li>
-                  <li>Decorators</li>
-                  <li>Enums</li>
-                  <li>Modules</li>
-                </ul>
-              </Card.Text>
-            </Card.Body>
-          </Card>
-        </Col>
-        <Col>
-          <Card bg="light" style={{ width: '18rem' }}>
-            <Card.Header>React</Card.Header>
-            <Card.Body>
-              <Card.Text>
-                <ul>
-                  <li>React Hooks</li>
-                  <li>React Router</li>
-                  <li>Redux</li>
-                  <li>Next.js</li>
-                  <li>Gatsby.js</li>
-                </ul>
-              </Card.Text>
-            </Card.Body>
-          </Card>
-        </Col>
-        <Col>
-          <Card bg="light" style={{ width: '18rem' }}>
-            <Card.Header>HTML</Card.Header>
-            <Card.Body>
-              <Card.Text>
-                <ul>
-                  <li>Semantic HTML</li>
-                  <li>Accessibility</li>
-                  <li>SEO</li>
-                  <li>Web Components</li>
-                  <li>Canvas</li>
-                </ul>
-              </Card.Text>
-            </Card.Body>
-          </Card>
-        </Col>
-        <Col>
-          <Card bg="light" style={{ width: '18rem' }}>
-            <Card.Header>CSS</Card.Header>
-            <Card.Body>
-              <Card.Text>
-                <ul>
-                  <li>CSS Grid</li>
-                  <li>CSS Flexbox</li>
-                  <li>SASS</li>
-                  <li>Bootstrap</li>
-                  <li>Material UI</li>
-                </ul>
-              </Card.Text>
-            </Card.Body>
-          </Card>
-        </Col>
+        {skills.map((skill, index) => (
+          <Col>
+            <SkillCard skill={skill} />
+          </Col>
+        ))
+        }
       </Row>
     </Container>
   );
 };
 
-export default Skills;
+export { Skills, SkillCard }
