+++ src/router/index.tsx
@@ -2,7 +2,7 @@ import { createHashRouter } from "react-router-dom";
 import App from '../App';
 import Radar from '../components/radar';
 import Home from '../components/home';
-import Skills from '../components/skills';
+import { Skills } from '../components/skills';
 
 export const router = createHashRouter([
   {
